requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pytesseract==0.3.10
PyMuPDF==1.23.8
openai==0.28.1
python-dotenv==1.0.0
pathlib2==2.3.7
gunicorn==21.2.0
requests==2.31.0
matplotlib==3.7.2
scikit-image==0.21.0
scipy==1.11.3